rename: 트레이닝 dto에 예약,리뷰 dto 폴더 생성 및 이동
