+++ src/main/java/com/fithub/fithubbackend/domain/Training/application/PaymentService.java
@@ -1,9 +1,9 @@
 package com.fithub.fithubbackend.domain.Training.application;
 
-import com.fithub.fithubbackend.domain.Training.dto.CancelReqDto;
-import com.fithub.fithubbackend.domain.Training.dto.PaymentReqDto;
-import com.fithub.fithubbackend.domain.Training.dto.PaymentResDto;
-import com.fithub.fithubbackend.domain.Training.dto.ReserveReqDto;
+import com.fithub.fithubbackend.domain.Training.dto.reservation.CancelReqDto;
+import com.fithub.fithubbackend.domain.Training.dto.reservation.PaymentReqDto;
+import com.fithub.fithubbackend.domain.Training.dto.reservation.PaymentResDto;
+import com.fithub.fithubbackend.domain.Training.dto.reservation.ReserveReqDto;
 import com.fithub.fithubbackend.domain.user.domain.User;
 import com.siot.IamportRestClient.exception.IamportResponseException;
