+++ src/main/java/com/fithub/fithubbackend/domain/Training/dto/reservation/PaymentResDto.java
@@ -1,4 +1,4 @@
-package com.fithub.fithubbackend.domain.Training.dto;
+package com.fithub.fithubbackend.domain.Training.dto.reservation;
 
 import lombok.Builder;
